requirements.txt
datasets
einops
optax
tqdm
pyseltongue
pycryptodome
